LangChain/requirements.txt
openai
chromadb
langchain
wikipedia
streamlit
tiktoken
python_dotenv